fix: fixed example issues on the new bottomsheet configuration
