--- Sources/BottomSheet/View Modifiers/SheetPlusModifiers.swift
@@ -1,12 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Wouter van de Kamp on 26/11/2022.
-//
-
-import Foundation
-
-extension SheetPlus {
-    
-}
